requirements.txt
aniso8601==7.0.0
asgiref==3.5.1
braintree==4.15.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.13
django-allauth==0.50.0
django-cors-headers==3.11.0
django-filter==21.1
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.13.1
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
idna==3.3
oauthlib==3.2.0
Pillow==9.1.0
promise==2.3
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
urllib3==1.26.9
whitenoise==6.0.0
